refactor: move server views
