+++ file_storage/web/api/router.py
@@ -2,7 +2,7 @@
 from fastapi.routing import APIRouter
 
 from file_storage.web.api import docs, echo, monitoring, user
-from file_storage.web.api.server.views import router as server_router
+from file_storage.web.server.views import router as server_router
 
 api_router = APIRouter()
 api_router.include_router(monitoring.router)
